+++ src/enum.ts
@@ -1,18 +1,16 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 
 export enum Commands {
-  TRANSFORM_TO_MOCK_SERVER_DB = "mockServer.transformToMockServerDB",
-  GENERATE_MOCK_FILES = "mockServer.generateMockFiles",
-  START_SERVER = "mockServer.startServer",
-  STOP_SERVER = "mockServer.stopServer",
-  RESET_SERVER = "mockServer.resetServer",
-  RESET_AND_RESTART = "mockServer.resetAndRestart",
-  START_WITH_NEW_PORT = "mockServer.startWithNewPort",
-  SWITCH_ENVIRONMENT = "mockServer.switchEnvironment",
-  GET_DB_SNAPSHOT = "mockServer.getDbSnapShot",
-  HOMEPAGE = "mockServer.homePage",
-  SET_PORT = "mockServer.setPort",
-  SET_ROOT = "mockServer.setRoot"
+  TRANSFORM_TO_MOCK_SERVER_DB = "mock-server.transformToMockServerDB",
+  GENERATE_MOCK_FILES = "mock-server.generateMockFiles",
+  START_SERVER = "mock-server.startServer", // Will Restart if already running. If Full Reload it will reset and restart.
+  STOP_SERVER = "mock-server.stopServer", // Will Reset
+  START_WITH_NEW_PORT = "mock-server.startWithNewPort",
+  SWITCH_ENVIRONMENT = "mock-server.switchEnvironment",
+  GET_DB_SNAPSHOT = "mock-server.getDbSnapShot",
+  OPEN_HOMEPAGE = "mock-server.openHomePage",
+  SET_PORT = "mock-server.setPort",
+  SET_ROOT = "mock-server.setRoot"
 }
 
 export enum ServerStatus {